remove all cr-controller references
