+++ .github/workflows/build.yaml
@@ -9,11 +9,11 @@ jobs:
     steps:
       - name: Checkout Repo
         uses: actions/checkout@v3
-      - name: Build Docker Image for site-manager-cr-controller
-        run: docker build -t site-manager -f site-manager-cr-controller/Dockerfile --no-cache site-manager-cr-controller
+      - name: Build Docker Image for site-manager
+        run: docker build -t site-manager --no-cache site-manager
       - name: Build Docker Image for sm-client
         run: docker build -t sm-client -f Dockerfile-sc --no-cache .
       - name: Build Docker Image for paas-geo-monitor
-        run: docker build -t paas-geo-monitor -f paas-geo-monitor/Dockerfile --no-cache paas-geo-monitor/
+        run: docker build -t paas-geo-monitor --no-cache paas-geo-monitor
       - name: Build Docker Image for sm-dummy
-        run: docker build -t sm-dummy -f tests/sm-dummy/Dockerfile --no-cache .
+        run: docker build -t sm-dummy --no-cache tests/sm-dummy
